2_news_research_tool_project/requirements.txt
langchain
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
OpenAI
PyPDF2